pyproject.toml
[project]
name = "x402-server"
version = "0.1.0"
description = "x402 backend server with CDP integration"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "cdp-sdk",
    "setuptools",
    "wheel",
    "ckzg>=2.0.0",
    "eth-account>=0.13.7",
    "x402",
    "httpx",
    "web3",  # 用于网络配置
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
